requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
certifi==2023.7.22
charset-normalizer==3.2.0
decorator==5.1.1
exceptiongroup==1.1.3
executing==2.0.0
greenlet==3.0.0
idna==3.4
ipython==8.16.1
jedi==0.19.1
matplotlib-inline==0.1.6
parso==0.8.3
password-strength==0.0.3.post2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.39
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
PyJWT==2.8.0
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.11.2
typing_extensions==4.8.0
urllib3==2.0.5
wcwidth==0.2.8
